Added many features

+++ src/main/java/dev/lightdream/bloodbath/dto/PluginLocation.java
@@ -8,8 +8,8 @@ import lombok.NoArgsConstructor;
 public class PluginLocation {
 
     public String world;
-    public int x;
-    public int y;
-    public int z;
+    public float x;
+    public float y;
+    public float z;
 
 }
